+++ masci_tools/cmdline/commands/__init__.py
@@ -1,4 +1,4 @@
 """
 Import all the command groups to include in the cli
 """
-from . import (fleur_schema, parse, plot, tools)
+from . import (convert_inpgen, fleur_schema, parse, plot, tools)